backend/user-service/pyproject.toml
[project]
name = "cashup-user-service"
version = "1.0.0"
description = "CashUp量化交易系统 - 用户服务"
requires-python = ">=3.9"
dependencies = [
    # 基础共享依赖
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.12.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    # 用户服务特有依赖
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "redis==5.0.1",
    "python-multipart==0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]